64-bit compilation works! Kinda. I guess. Everything runs now after several bugfixes, but I really don't know if the traditional VM stuff would still work or if it's a complete mess now. Then again, Jomme in its current state wouldn't work as a qvm anyway so maybe it doesn't matter. But yeah... dirty dirty. :)

+++ CODE-mp/ghoul2/G2_misc.cpp
@@ -1517,7 +1517,7 @@ void *G2_FindSurface(void *mod_t, int index, int lod)
 	model_t	*mod = (model_t *)mod_t;
 
 	// point at first lod list
-	byte	*current = (byte*)((int)mod->mdxm + (int)mod->mdxm->ofsLODs);
+	byte	*current = (byte*)((size_t)mod->mdxm + (size_t)mod->mdxm->ofsLODs);
 	int i;
 
 	//walk the lods
